rename: weather entities의 consts 파일 구조 변경
